refactor: 정적 메소드의 파라미터 타입 형변환 처리 없이 받도록 수정

+++ src/main/java/kitchenpos/dto/menu/MenuRequest.java
@@ -3,26 +3,25 @@ package kitchenpos.dto.menu;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public class MenuRequest {
     private String name;
-    private BigDecimal price;
+    private long price;
     private Long menuGroupId;
     private List<MenuProductRequest> menuProducts;
 
     protected MenuRequest() {
     }
 
-    public MenuRequest(final String name, final BigDecimal price, final Long menuGroupId, final List<MenuProductRequest> menuProducts) {
+    public MenuRequest(final String name, final long price, final Long menuGroupId, final List<MenuProductRequest> menuProducts) {
         this.name = name;
         this.price = price;
         this.menuGroupId = menuGroupId;
         this.menuProducts = menuProducts;
     }
 
-    public static MenuRequest of(final String name, final BigDecimal price, final Long menuGroupId, final List<MenuProductRequest> menuProducts) {
+    public static MenuRequest of(final String name, final long price, final Long menuGroupId, final List<MenuProductRequest> menuProducts) {
         return new MenuRequest(name, price, menuGroupId, menuProducts);
     }
 
@@ -34,7 +33,7 @@ public class MenuRequest {
         return this.name;
     }
 
-    public BigDecimal getPrice() {
+    public long getPrice() {
         return this.price;
     }
